arreglandeo el error del cron

+++ app/Http/Controllers/CronController.php
@@ -106,7 +106,9 @@ class CronController extends Controller
                         $this->jellyfin->setCredentials($cjf->jellyfinserver);
                         $user = json_decode($cjf->json_data, true);
                         if($user){
-                            $this->jellyfin->provider->deleteUser($user['Id']);
+                            if(array_key_exists('Id', $user)){
+                                $this->jellyfin->provider->deleteUser($user['Id']);
+                            }
                             $cjf->status = false;
                             $cjf->save();
                             $contadorCJS++;
@@ -128,6 +130,9 @@ class CronController extends Controller
                         $this->jellyfin->setCredentials($djf->jellyfinserver);
                         $user = json_decode($djf->json_data, true);
                         if($user){
+                            if(array_key_exists('Id', $user)){
+                                $this->jellyfin->provider->deleteUser($user['Id']);
+                            }
                             $this->jellyfin->provider->deleteUser($user['Id']);
                             $djf->delete();
                             $contadorDJS++;
